requirements.txt
streamlit
groq
langchain
langchain_community
PyPDF2
langchain-groq
ollama
python-dotenv